requirements.txt
flask
Flask-SQLAlchemy
flask-login
Flask-CORS
gunicorn
psycopg2
requests
python-dotenv
urllib3